+++ src/main/java/ar/edu/itba/pf/domain/environment/windengine/WindStrategy.java
@@ -5,7 +5,7 @@ package ar.edu.itba.pf.domain.environment.windengine;
  */
 public interface WindStrategy {
 
-    double getX(int x, int y, int t);
+    double getX(int x, int y, long t);
 
-    double getY(int x, int y, int t);
+    double getY(int x, int y, long t);
 }
