+++ src/main/java/ar/edu/itba/pf/domain/environment/windengine/impl/PolarWind.java
@@ -2,7 +2,6 @@ package ar.edu.itba.pf.domain.environment.windengine.impl;
 
 import ar.edu.itba.pf.domain.environment.windengine.WindStrategy;
 
-import static java.lang.Math.PI;
 import static java.lang.Math.cos;
 import static java.lang.Math.sin;
 
@@ -17,12 +16,12 @@ public class PolarWind implements WindStrategy {
     }
 
     @Override
-    public double getX(int x, int y, int t) {
+    public double getX(int x, int y, long t) {
         return modulus*cos(angle);
     }
 
     @Override
-    public double getY(int x, int y, int t) {
+    public double getY(int x, int y, long t) {
         return modulus*sin(angle);
     }
 }
